# documentatie/openAPI.yaml
openapi: 3.0.0
info:
  title: Opdrachtenbox API
  description: API voor het beheren van opdrachten voor studenten van de opleiding CMD
  version: 0.0.3
servers:
  - url: http://localhost:8080/
    description: Local development server
...
paths:
...
  /business:
    put:
      tags:
        - Bedrijven
      summary: updaten van bedrijf
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/BusinessDto"
      description: "Alleen toegankelijk voor: opdrachtgevers van het bedrijf"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: bedrijf successful aangepast
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker is niet geauthoriseerd om de pagina van bedrijf aan te passen
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /business/{businessId}:
    get:
      tags:
        - Bedrijven
      parameters:
        - name: businessId
          in: path
          required: true
          description: id van het bedrijf
          schema:
            type: integer
            format: int32
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: bedrijf successful opgehaald
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /projects:
    get:
      tags:
        - Projecten
      summary: alle projecten ophalen
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      parameters:
        - in: query
          name: businessId
          schema:
            type: integer
            format: int32
          required: false
      responses:
        "200":
          description: projecten succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetProjectDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "5XX":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Projecten
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 50
                  example: Nieuwe project titel
                description:
                  type: string
                  maxLength: 400
                  example: Nieuwe project beschrijving
                image:
                  type: string
                  format: binary
                  description: afbeelding van het project
      summary: nieuw project aanmaken
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: project succesvol aangemaakt
          content:
            application/json:
              schema:
                type: integer
                format: int32
                example: 25
                description: Het id van het aangemaakte project
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "5XX":
          $ref: "#/components/responses/ServerError"
  /projects/all:
    get:
      tags:
        - Projecten
      summary: ophalen van alle projecten
      description: "Toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: projecten succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessProjectsWithTasksAndSkillsDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de docenten rol
        "5XX":
          $ref: "#/components/responses/ServerError"
  /projects/{projectId}:
    get:
      tags:
        - Projecten
      summary: Een enkel project ophalen
      parameters:
        - name: projectId
          in: path
          required: true
          description: de id van het project
          schema:
            type: integer
            format: int32
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: Project succesvol opgehaald
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProjectDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "404":
          description: project heeft een foute opbouw
        "5XX":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
  responses:
    BadRequest:
      description: request body heeft een foute opbouw
    UnAuthenticated:
      description: gebruiker is niet ingelogd
    ServerError:
      description: server fout
  schemas:
    GetProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
        title:
          type: string
          maxLength: 50
          example: Project titel
        description:
          type: string
          maxLength: 400
          example: Project beschrijving
        projectTopSkills:
          type: array
          items:
            $ref: "#/components/schemas/GetSkillDto"
        business:
          $ref: "#/components/schemas/BusinessDto"
        photo:
          $ref: "#/components/schemas/FileDto"
      required:
        - projectId
        - title
        - description
        - business
        - projectTopSkills
    BusinessProjectsWithTasksAndSkillsDto:
      type: object
      properties:
        business:
          $ref: "#/components/schemas/BusinessDto"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectWithTasksAndSkillsDto"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/GetSkillDto"
    ProjectWithTasksAndSkillsDto:
      type: object
      properties:
        projectId:
          type: integer
          format: int32
          example: 25
        title:
          type: string
          maxLength: 50
          example: Project titel
        description:
          type: string
          maxLength: 400
          example: Project beschrijving
        image:
          $ref: "#/components/schemas/FileDto"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskWithSkills"
    TaskWithSkills:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
          example: 12
        title:
          type: string
          maxLength: 50
          example: Taak titel
        description:
          type: string
          maxLength: 400
          example: Taak beschrijving
        totalNeeded:
          type: integer
          format: int32
          example: 3
        totalAccepted:
          type: integer
          format: int32
          example: 1
        totalRegistered:
          type: integer
          format: int32
          example: 2
        skills:
          type: array
          items:
            $ref: "#/components/schemas/GetSkillDto"
...
    GetSkillDto:
      description: een skill object
      type: object
      properties:
        skillId:
          type: integer
          format: int32
          example: 20
        name:
          type: string
          maxLength: 50
          example: adobe premiere pro
        isPending:
          type: boolean
          example: false
      required:
        - skillId
        - name
        - isPending
...
    FileDto:
      type: object
      properties:
        path:
          type: string
          example: /image.png
          maxLength: 255
...
    BusinessDto:
      type: object
      properties:
        businessId:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          maxLength: 255
          example: mijn naam
        description:
          type: string
          maxLength: 4000
          example: Dit bedrijf is een bedrijf dat ...
        photo:
          $ref: "#/components/schemas/FileDto"
        location:
          type: string
          maxLength: 255
      required:
        - name
        - description
        - photo
        - location
...